+++ src/Validation/dtos/index.ts
@@ -20,3 +20,4 @@ export { IValidationIsjwtDto } from "./validation.isjwt.dto";
 export { IValidationIsStrongPasswordDto } from "./validation.isstrongpassword.dto";
 export { IValidationIsEmailDto } from "./validation.isemail.dto";
 export { IValidationSameDto } from "./validation.same.dto";
+export { IValidationRequiredWithDto } from "./validation.requiredwith.dto";