+++ src/Validation/dtos/validation.requiredwith.dto.ts
@@ -1,0 +1,7 @@
+import { ValidationGeneralDto } from "./validation.general.dto";
+
+export interface IValidationRequiredWithDto extends Omit<ValidationGeneralDto, 'params' | 'customFunction'> {
+    params: {
+        fields: string[]
+    }
+}